requirements.txt
Flask==2.0.1
pandas==1.3.0
openpyxl==3.0.7
xlrd==2.0.1
matplotlib==3.4.2
seaborn==0.11.1
reportlab==3.6.1
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1
pytest==6.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==0.15.1
itsdangerous==2.0.1
email-validator==1.1.3
bcrypt==3.2.0
